listener: bugfix: handle warnings around linenoise.
* parser.c (repl): Also push and pop the warning handler
around the linenoise call, like we do around evaluation.
The reason is that when the library is used in source code
form, it can generate warnings while loading. Loading can
be triggered by completion from inside linenoise. Neglecting
to handle warnings and throw a continue causes a problem
because the warnings then fall victim to the listener's
master catch. That then interferes with the load.