for/for*: stricter syntax check.
* eval.c (me_for): Require at least one argument.
However, we let the init-forms continue to be optional
and document it.

* txr.1: Refer to for and for* as macros, since they have been
since 2016. The omission of the inc-form list is shown
as a second variant of the syntax. This is to avoid misleading
the reader into thinking that the the inc-form list can be
omitted while body forms are present. A spurious paragraph
reiterating that the macros establish an anonymous block is
removed. That extra text was present in the first draft written
in 2011, and maintained since.

* stdlib/doc-syms.tl: Updated.