syrooting: remove two useless checks.
* txr.c (sysroot_init): The value of TXR_REL_PATH
already has the exe suffix on it if we are building
for Windows. There is no need to try with EXE_SUFF
tacked on to it. There is also no need to try PROG_NAME
with or without the suffix. EXE_SUFF is the empty
string on non-Windows platforms, so the two lines of
code are equivalent. On Windows, PROG_NAME without
the EXE_SUFF will not match; that maybe_sysroot call
is a waste of CPU cycles.