New function: search-all
* eval.c (eval_init): search-all intrinsic registered.

* lib.c (search_common): New Boolean argument all,
indicating whether all positions are to be returned.
We must handle this in the two places where empty
key and sequence are handled, and also in the main loop.
A trick is used: the found variable is now bound by
list_collect_decl, but not used for collecting unless
all is true.
(search, rsearch, contains): Pass 0 for all argument
of search_common.
(search_all): New function.

* lib.h (search_all): Declared.

* tests/012/seq.tl: New tests.

* txr.1: Documented.

* stdlib/doc-syms.tl: Regenerated.