put-json: restore indent on unwinding.
The following situation is observed in the listener.

1> (put-json #H(() (a 1)))
{** print: invalid object a in JSON
 ** during evaluation at expr-1:1 of form (put-json #H(() (a 1)))
 1>
 1>

An indent established in the aborted JSON print job
has been left in the stream.

* lib.c (put_json): Save the indent value also, not
only the mode. Restore the indent mode and value
on unwinding, not just on a normal exit from out_json_rec,
similiarly to what obj_print does.