New feature: struct preludes.
A struct prelude definition associates one or more
future defstruct (by struct name) with clauses which
are implicitly inserted into the defstruct.

It is purely a macro-time construct, customizing the
expansion behavior of defstruct.

* stdlib/struct.tl (*struct-prelude, *struct-prelude-alists*):
New special variables holding hash tables.
(defstruct): Before processing slot-specs, augment it with
the contents of the prelude definitions associated with
this struct name.
(define-struct-prelude): New macro.

* autoload.c (struct_set_entries): define-struct-prelude
is interned and triggers autoload of struct module.

* tests/012/oop-prelude.tl: New file.

* tests/012/oop-prelude.expected: Likewise.

* txr.1: Documented.

* stdlib/doc-syms.tl: Updated.