fix --no-gen-gc configuration.
This fixes only the build. I'm getting a crash in
one test case, namely tests/010/json.tl.

* lib.h (mut): Remove stray semicolon from definition.
This semicolon compensates for the lack of a semicolon
in txr.c, which becomes a syntax errror under no-gen-gc,
when the other definition of mut is active.
(mkloc, setcheck): Let's add casts of the object argument
to void.  This gets rid of a number of unused parameter
errors in various functions that take an object parameter
that is only used in the case of generational GC.

* txr.c (txr_main): Add missing semicolon after mut call.

* gc.c (gc_wrap): In the no CONFIG_GEN_GC case, cast
argument full to void, since it is unused.