bugfix: gc issue in parser interface.
* parser.c (lisp_parse_impl): Add a call to gc_hint,
to protect the parser object from garbage collection.
We allocate this object, and then use a raw pointer to
the parser, which leaves it exposed. This resulted in
a crash of the tests/010/json.tl test case (which runs
with the --gc-debug mode). The crash doesn't occur in
a regular build; it reproduced in build configured
with --no-gen-gc. Possibly, why it doesn't repro under
generational GC is that the hash table which associates
streams and parsers may be moving the objects into the
mature generation.