txr: bugfix, allow lazy lists in multi match.
* match.c (do_match_line): Handle LCONS the same way as CONS.
When a variable occurs whose value is a list of strings,
that may be a lazy list. I ran into a problem using @(data x)
to capture a list of strings, and then matching that with @x;
the error being "unsupported object in spec", caused by the
list's LCONS type not handled in this switch.