match: support @nil in predicates.
For instance @(<= 10 @nil 20) is a pattern which matches
a number between 10 and 20, without binding a variable.

* stdlib/match.tl (compile-predicate-match): Looks like
this code was already halfway expressing the intent that
the avar could be nil, because arg-var takes the value
of avar if that is non-nil, otherwise a gensym is
substituted. What was missing was that the gensym that
replaces nil must also be substituted into the predicate.

* tests/011/patmatch.tl: New tests.

* txr.1: Document that the variable embedded in a
predicate may be null.