autoload: temporarily disable backward compat.
The purpose of this change is to help with the situation
when an uncompiled stdlib is being used (.tl files, no .tlo)
and txr is invoked with -C <num> to select compatibility
mode with an old version. The problem with compatibility is
that it potentially breaks the library due to the different
behavior of some macros like caseql.

Some test cases in the test suite use backwards compatibility,
and sometimes it is necessary to run with the uncompiled
test suite when debugging compiler work: situations when the
compiler is too broken to build the library.

* autoload.c (autoload_try): Temporarly set the opt_compat
option to 0 (disabled) around autoload processing. Thus
the loading of library code in source code form will not be
adversely affected by any syntax or macro level backward
compatibility hacks.